refactored cast options
