+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuApproximatePercentile.scala
@@ -144,9 +144,7 @@ case class ApproxPercentileFromTDigestExpr(
           // array and return that (after converting from Double to finalDataType)
           withResource(cv.getBase.approxPercentile(Array(p))) { percentiles =>
             withResource(percentiles.extractListElement(0)) { childView =>
-              withResource(doCast(childView, DataTypes.DoubleType, finalDataType,
-                  ansiMode = false, legacyCastToString = false,
-                  stringToDateAnsiModeEnabled = false)) { childCv =>
+              withResource(doCast(childView, DataTypes.DoubleType, finalDataType)) { childCv =>
                 GpuColumnVector.from(childCv.copyToColumnVector(), dataType)
               }
             }
@@ -159,9 +157,7 @@ case class ApproxPercentileFromTDigestExpr(
               GpuColumnVector.from(percentiles.incRefCount(), dataType)
             } else {
               withResource(percentiles.getChildColumnView(0)) { childView =>
-                withResource(doCast(childView, DataTypes.DoubleType, finalDataType,
-                    ansiMode = false, legacyCastToString = false,
-                    stringToDateAnsiModeEnabled = false)) { childCv =>
+                withResource(doCast(childView, DataTypes.DoubleType, finalDataType)) { childCv =>
                   withResource(percentiles.replaceListChild(childCv)) { x =>
                     GpuColumnVector.from(x.copyToColumnVector(), dataType)
                   }
